extract SettingsView into own file
also fix some copyright attributions
